requirements.txt
fastapi
uvicorn
motor
pydantic
bcrypt
PyJWT
cloudinary
python-multipart